Added Category and placeholders.
Tabs contains and lists categories.

The backend isn't set correctly. That still needs research and time to get sorted. but for now I have a loop around. (it has taken a week of research but I have no clue what the name is for what I am looking for. All I know is that it's a parent/child relationship thing)
The frontend has a temporary layout and theme. I just need this to not only help get the idea of how I can colour it, but to all be able to visually see my work coming together.

Next step is to fix the front end to sit better with the list of category along with colour themes.